Reservation 2nd Tier Done

+++ cCarRentalSystem/Client/Domain/Models/Reservation.cs
@@ -7,8 +7,8 @@ public class Reservation
     public int CarId{get;set;}
     public DateTime StartDate{get;set;}
     public DateTime EndDate{get;set;}
-    public int TotalPrice{get;set;}
-    public Reservation(int id,int userId, int carId, DateTime startDate, DateTime endDate, int totalPrice)
+    public double TotalPrice{get;set;}
+    public Reservation(int id,int userId, int carId, DateTime startDate, DateTime endDate, double totalPrice)
     {
         Id = id;
         UserId = userId;
